Organize JS modules under 150 lines
