+++ js/main.js
@@ -1,6 +1,6 @@
 import './core/index.js';
-import './builder.js';
+import './builder/index.js';
 import './ui.js';
 import './player/index.js';
 import './tests.js';
-import './worldgen.js';
+import './world/index.js';
