+++ js/player/movement/index.js
@@ -1,0 +1,2 @@
+import './input.js';
+import './loop.js';
